+++ src/NodeUtils.ts
@@ -152,6 +152,7 @@ export class NodeUtils {
                 }
 
                 node['parentNode'] = value;
+                node['obfuscated'] = false;
             }
         });
     }