+++ src/interfaces/nodes/INode.d.ts
@@ -1,4 +1,5 @@
 export interface INode {
     type: string;
     parentNode?: INode;
+    obfuscated?: boolean;
 }