+++ webpack.config.js
@@ -1,15 +1,13 @@
 'use strict';
 
-let fs = require("fs"),
+var fs = require("fs"),
     nodeExternals = require('webpack-node-externals'),
-    webpack = require('webpack');
+    webpack = require('webpack'),
+    OptimizeJSPlugin = require('optimize-js-plugin');
 
 function getLicenseText () {
-    return `/*
-Copyright (C) 2016 Timofey Kachalov <sanex3339@yandex.ru>
-
-${fs.readFileSync('./LICENSE.BSD', 'utf8')}
-*/`;
+    return "/*\nCopyright (C) 2016 Timofey Kachalov <sanex3339@yandex.ru>\n\n" +
+        fs.readFileSync('./LICENSE.BSD', 'utf8') + "\n*/";
 }
 
 module.exports = {
@@ -25,17 +23,16 @@ module.exports = {
         ]
     },
     resolve: {
-        extensions: ['', '.ts'],
-        modulesDirectories: ['./src', './node_modules']
+        extensions: ['.ts']
     },
     plugins: [
         new webpack.BannerPlugin(
             {
-                banner: `${getLicenseText()}\n\nrequire("source-map-support").install();\n`,
+                banner: getLicenseText() + '\n\nrequire("source-map-support").install();\n',
                 raw: true,
                 entryOnly: false
             }
-        )
+        ),
     ],
     output: {
         path: './dist',